pyproject.toml
[project]
name = "document-scanner"
version = "0.1.0"
description = "Hotfolder-based document processing pipeline"
authors = [{name = "AutoGenerated"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=4.0.0",
    "PyYAML>=6.0",
    "pdfplumber>=0.11.0",
    "pypdf>=4.0.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "openai>=1.51.0",
    "python-dotenv>=1.0.1",
    "jsonschema>=4.23.0",
    "requests>=2.32.3",
    "msal>=1.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
